Adding most of the TFC api.
The TFC API is kind of broken, so I deleted all classes that depend on
TFC common code to get this to build at all. Also, the code is in the
src/api/java, so it doesn't get packaged in the jar. so no actual
conflicts with TFC at runtime.