Fixes #3
Changed the implementation to ASM in custom behaviour instead of
reflecting in TFC water in the FluidRegistry. The last changes from
KingLemming has made it impossible to have the same fluid with two keys
in the FluidRegistry. The ASM will try to intercept requests for water,
and return TFC freshwater instead.